[BUGFIX] Only set PHPMyAdmin cookie for BE Users
When a frontend user logs in, then the
"BeUserAuthLogOffHook" creates a phpmyadmin cookie
and thus, exposes the usage of EXT:phpmyadmin for website
users even if they never log in to the TYPO3 Backend.

This change adapts the code to only work with BE_USER
log offs, and not frontend users.